Added tool test: PartialCharset findIndexOrAdd full.

+++ tools/partial-charset.h
@@ -6,7 +6,8 @@
 typedef struct PartialCharset
 {
 	Charset charset;
-	uint8_t numUsedChars;
+	// uint8_t is too small to store a count of 256.
+	uint16_t numUsedChars;
 } PartialCharset;
 
 int partialCharsetFindIndexOrAdd(PartialCharset *charset, CharsetChar charsetChar);
